+++ functions/Config.fnc.php
@@ -94,3 +94,51 @@ function ProgramConfig( $program, $item = 'all'  )
 	else
 		return $_ROSARIO['ProgramConfig'][ (string) $program ][ (string) $item ][1]['VALUE'];
 }
+
+
+
+/**
+ * Program User Config
+ * To get all config options at once
+ * If you want only one option, prefer `Preferences()`
+ *
+ * @example $gradebook_config = ProgramUserConfig( 'Gradebook' );
+ *
+ * @see Preferences()
+ * @see PROGRAM_USER_CONFIG table
+ *
+ * @since 2.9
+ *
+ * @param string  $program  Gradebook|WidgetsSearch|StaffWidgetsSearch|
+ * @param integer $staff_id Staff ID (optional). Defaults to User( 'STAFF_ID' ).
+ *
+ * @return array Program User Config, associative array( '[title]' => '[value]' ).
+ */
+function ProgramUserConfig( $program, $staff_id = 0 )
+{
+	static $program_config;
+
+	if ( ! $program )
+	{
+		return array();
+	}
+
+	$staff_id = $staff_id ? $staff_id : User( 'STAFF_ID' );
+
+	$config_RET = DBGet( DBQuery( "SELECT TITLE,VALUE
+		FROM PROGRAM_USER_CONFIG
+		WHERE USER_ID='" . $staff_id . "'
+		AND PROGRAM='" . $program . "'" ), array(), array( 'TITLE' ) );
+
+	if ( $config_RET )
+	{
+		foreach ( (array) $config_RET as $title => $value )
+		{
+			$program_config[ $program ][ $staff_id ][ $title ] = $value[1]['VALUE'];
+		}
+	}
+	else
+		$program_config[ $program ][ $staff_id ] = null;
+
+	return $program_config[ $program ][ $staff_id ];
+}