+++ functions/StaffWidgets.fnc.php
@@ -114,18 +114,25 @@ function StaffWidgets( $item, &$myextra = null )
 		// User Widgets (configured in My Preferences).
 		case 'user':
 
-			$widgets_RET = DBGet( DBQuery( "SELECT TITLE
+			/*$widgets_RET = DBGet( DBQuery( "SELECT TITLE
 				FROM PROGRAM_USER_CONFIG
 				WHERE USER_ID='" . User( 'STAFF_ID' ) . "'
 				AND PROGRAM='StaffWidgetsSearch'" .
 				( count( $_ROSARIO['StaffWidgets'] ) ?
 					"AND TITLE NOT IN ('" . implode( "','", array_keys( $_ROSARIO['StaffWidgets'] ) ) . "')" :
 					''
-				) ) );
+				) ) );*/
 
-			foreach ( (array) $widgets_RET as $widget )
+			$user_widgets = ProgramUserConfig( 'StaffWidgetsSearch' );
+
+			$saved_widget_titles = array_keys( $_ROSARIO['StaffWidgets'] );
+
+			foreach ( (array) $user_widgets as $user_widget_title => $value )
 			{
-				StaffWidgets( $widget['TITLE'], $extra );
+				if ( ! in_array( $user_widget_title, $saved_widget_titles ) )
+				{
+					StaffWidgets( $user_widget_title, $extra );
+				}
 			}
 
 		break;