+++ modules/Grades/InputFinalGrades.php
@@ -118,13 +118,7 @@ if ( $_REQUEST['modfunc']=='gradebook')
 {
 	if ( $_REQUEST['mp'])
 	{
-		$config_RET = DBGet(DBQuery("SELECT TITLE,VALUE FROM PROGRAM_USER_CONFIG WHERE USER_ID='".User('STAFF_ID')."' AND PROGRAM='Gradebook'"),array(),array('TITLE'));
-
-		if (count($config_RET))
-			foreach ( (array) $config_RET as $title => $value)
-				$programconfig[User('STAFF_ID')][ $title ] = $value[1]['VALUE'];
-		else
-			$programconfig[User('STAFF_ID')] = true;
+		$gradebook_config = ProgramUserConfig( 'Gradebook' );
 
 		$_ROSARIO['_makeLetterGrade']['courses'][ $course_period_id ] = DBGet(DBQuery("SELECT DOES_BREAKOFF,GRADE_SCALE_ID FROM COURSE_PERIODS WHERE COURSE_PERIOD_ID='".$course_period_id."'"));
 
@@ -153,10 +147,10 @@ if ( $_REQUEST['modfunc']=='gradebook')
 					$total = $total_percent = 0;
 
 					foreach ( (array) $student as $partial_points)
-						if ( $partial_points['PARTIAL_TOTAL']!=0 || $programconfig[User('STAFF_ID')]['WEIGHT']!='Y')
+						if ( $partial_points['PARTIAL_TOTAL']!=0 || $gradebook_config['WEIGHT']!='Y')
 						{
-							$total += $partial_points['PARTIAL_POINTS']*($programconfig[User('STAFF_ID')]['WEIGHT']=='Y'?$partial_points['FINAL_GRADE_PERCENT']/$partial_points['PARTIAL_TOTAL']:1);
-							$total_percent += ($programconfig[User('STAFF_ID')]['WEIGHT']=='Y'?$partial_points['FINAL_GRADE_PERCENT']:$partial_points['PARTIAL_TOTAL']);
+							$total += $partial_points['PARTIAL_POINTS']*($gradebook_config['WEIGHT']=='Y'?$partial_points['FINAL_GRADE_PERCENT']/$partial_points['PARTIAL_TOTAL']:1);
+							$total_percent += ($gradebook_config['WEIGHT']=='Y'?$partial_points['FINAL_GRADE_PERCENT']:$partial_points['PARTIAL_TOTAL']);
 						}
 
 					if ( $total_percent!=0)
@@ -216,8 +210,8 @@ if ( $_REQUEST['modfunc']=='gradebook')
 
 				foreach ( (array) $percents as $percent)
 				{
-					$total += $percent['GRADE_PERCENT'] * $programconfig[User('STAFF_ID')][$prefix.$percent['MARKING_PERIOD_ID']];
-					$total_percent += $programconfig[User('STAFF_ID')][$prefix.$percent['MARKING_PERIOD_ID']];
+					$total += $percent['GRADE_PERCENT'] * $gradebook_config[$prefix.$percent['MARKING_PERIOD_ID']];
+					$total_percent += $gradebook_config[$prefix.$percent['MARKING_PERIOD_ID']];
 				}
 				$total /= $total_percent;